requirements.txt
python-telegram-bot==20.7
langchain==0.1.9
langchain-community>=0.0.21
chromadb==0.4.22
sentence-transformers==2.3.1
pypdf>=4.0.0
python-docx==1.1.0
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.6.1
aiohttp==3.9.1
